Fix concurrency issue in `SampleActivity`
BillingServiceStubImpl makes use of an injected instance of SkuDetailsBundleBuilder. SkuDetailsBundleBuilder exposes a `newBuilder` method, that by design does not create a new instance of the class but rather just resets its internal fields. This makes it impossible to query for different product types at the same time. For now, we need to serialize access to that object by fetching the SKU details sequentially. Long term speaking, we need to redesign this part of the code to allow for truly parallel access.