Adapt Register to using the new Google Play Billing library (#49)
* Update tooling versions and target SDK

* Fix existing implementation of `BillingServiceStubImpl`

The API version of the current AIDL interface is 5, but potentially the service could be queried for a higher version.

Co-authored-by: Eugenio Marletti <takhion@gmail.com>

* Add implementation of the new `GoogleServiceProvider`

Co-authored-by: Eugenio Marletti <takhion@gmail.com>

* Remove redundant AIDL file for In-app Billing

* Adapt sample app to using PBL, remove obsolete classes and tests

* Fix JSON parsing (`itemType` -> `type`)

Co-authored-by: Eugenio Marletti <takhion@gmail.com>

* Update Travis config file to accept new license agreement

* Break line down to adhere to checkstyle guidelines

* Fix a bunch of style, move two static fields within appropriate class

* Fix keys for requesting purchases

* Fix concurrency issue in `SampleActivity`

BillingServiceStubImpl makes use of an injected instance of SkuDetailsBundleBuilder. SkuDetailsBundleBuilder exposes a `newBuilder` method, that by design does not create a new instance of the class but rather just resets its internal fields. This makes it impossible to query for different product types at the same time. For now, we need to serialize access to that object by fetching the SKU details sequentially. Long term speaking, we need to redesign this part of the code to allow for truly parallel access.