Move to Single Activity
Remove Settings/Buy/Legal activities; navigate to corresponding Fragments from MainActivity.

Changes the theme of the MainActivity so that it's by default translucent (to keep the illusion when a user is buying something). This change of theme also changes the windowBackground of the Fragments, just so that they all have the same colors as before this change (note: the BuyFragment is a dark theme now just because it would be a larger change to also tweak all of its visuals to match as before. Larger changes are slated in #83)